Remove the reflected binary arithmetic operations
The reference documentation states that the reflected operations
(e.g., __rsub__) are "only called if the left operand does not support
the corresponding operation and the operands are of different types".
As we do not wish to support such operations for classes other than
ProcSet, it is unnecessary to implement the reflected operations.